Add support for b3 headers and reorganize
