+++ src/proxy/http/mod.rs
@@ -22,7 +22,7 @@ pub mod router;
 pub mod settings;
 pub mod strip_header;
 pub mod timeout;
-pub mod trace;
+pub mod trace_context;
 pub mod upgrade;
 
 pub use self::client::Client;
